Rename useUserV2 to useUser
