+++ next/pages/moj-profil.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'next-i18next'
 import AccountSectionHeader from '../components/forms/segments/AccountSectionHeader/AccountSectionHeader'
 import AccountPageLayout from '../components/layouts/AccountPageLayout'
 import { SsrAuthProviderHOC } from '../components/logic/SsrAuthContext'
-import { prefetchUserQuery } from '../frontend/hooks/useUserV2'
+import { prefetchUserQuery } from '../frontend/hooks/useUser'
 import { amplifyGetServerSideProps } from '../frontend/utils/amplifyServer'
 import { slovakServerSideTranslations } from '../frontend/utils/slovakServerSideTranslations'
